Perfected Floyd Steinberg Ditherer

+++ Project1/ColorQuantizer/Logic/Quantizers/IQuantizer.cs
@@ -8,9 +8,9 @@ namespace ImageProcessing.Logic.Quantizers
     public interface IQuantizer
     {
         void AddColor(System.Drawing.Color c);
-        byte GetPaletteIndex(Color c);
-        byte GetPaletteIndex(System.Drawing.Color c);
+        int GetPaletteIndex(Color c);
+        int GetPaletteIndex(System.Drawing.Color c);
         List<Color> GetPalette();
-        Color GetColorByIndex(byte index);
+        Color GetColorByIndex(int index);
     }
 }
